Feature/frontend

+++ backend/src/main/java/com/sdacademy/budgettracker/service/UserService.java
@@ -11,7 +11,7 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    public User getCurrentUser(Double userId) {
+    public User getCurrentUser(Long userId) {
         return userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
     }
